+++ backend/parallel.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: parallel.c,v 1.35 2002/01/27 21:16:08 mike Exp $"
+ * "$Id: parallel.c,v 1.36 2002/02/12 19:05:36 mike Exp $"
  *
  *   Parallel port backend for the Common UNIX Printing System (CUPS).
  *
@@ -261,7 +261,8 @@ main(int  argc,		/* I - Number of command-line arguments (6 or 7) */
       }
 
       if (argc > 6)
-	fprintf(stderr, "INFO: Sending print file, %u bytes...\n", tbytes);
+	fprintf(stderr, "INFO: Sending print file, %lu bytes...\n",
+	        (unsigned long)tbytes);
     }
   }
 
@@ -639,5 +640,5 @@ list_devices(void)
 
 
 /*
- * End of "$Id: parallel.c,v 1.35 2002/01/27 21:16:08 mike Exp $".
+ * End of "$Id: parallel.c,v 1.36 2002/02/12 19:05:36 mike Exp $".
  */
