+++ backend/usb.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: usb.c,v 1.23 2002/01/27 21:16:09 mike Exp $"
+ * "$Id: usb.c,v 1.24 2002/02/12 19:05:37 mike Exp $"
  *
  *   USB port backend for the Common UNIX Printing System (CUPS).
  *
@@ -250,7 +250,8 @@ main(int  argc,		/* I - Number of command-line arguments (6 or 7) */
       }
 
       if (argc > 6)
-	fprintf(stderr, "INFO: Sending print file, %u bytes...\n", tbytes);
+	fprintf(stderr, "INFO: Sending print file, %lu bytes...\n",
+	        (unsigned long)tbytes);
     }
   }
 
@@ -450,5 +451,5 @@ list_devices(void)
 
 
 /*
- * End of "$Id: usb.c,v 1.23 2002/01/27 21:16:09 mike Exp $".
+ * End of "$Id: usb.c,v 1.24 2002/02/12 19:05:37 mike Exp $".
  */
